모듈화 적용
