+++ src/api/axiosApi/applicantApi.js
@@ -1,4 +1,3 @@
 export const Applicant = {
-    
+  ApplicantSearchList: "/api/manage-user/applicantListBody.do",
 };
-
