+++ src/components/hook/applicant/useApplicantListSearchQuery.js
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/vue-query";
+import { applicantListSearchApi } from "../../../api/applicant/applicantListSearchApi";
+
+export const useApplicantListSearchQuery = (injectedValue, cPage) => {
+  return useQuery({
+    queryKey: ["applicantList", injectedValue, cPage],
+    queryFn: () => applicantListSearchApi(injectedValue, cPage),
+  });
+};
